v0.2.0
- Fix issue #3
- Update deps
