+++ docs/app.js
@@ -1,1 +1,1 @@
-!function(){const e=e=>e.map(e=>[[...e],[...e.map(e=>e.toUpperCase())]]).reduce((e,t)=>[...e,...t],[]),t="bcdđhgklmnpqrstvx",n=[t,t.toUpperCase()].join(""),r=["aăâeêioôơuưy","áắấéếíóốớúứý","àằầèềìòồờùừỳ","ảẳẩẻểỉỏổởủửỷ","ãẵẫẽễĩõỗỡũữỹ","ạặậẹệịọộợụựỵ"],o=[...r].map(e=>e+e.toUpperCase()).map(e=>e.split("")).reduce((e,t)=>[...e,...t],[]).join(""),s=[...r].map(e=>[e[1],e[2],e[4],e[7],e[8],e[10]]).reduce((e,t)=>[...e,...t],[]).join(""),i=[...r].slice(1).join(""),c=["ab,ace,act,ad,ae,ann,al,ag,ar,as","bb,br,bl","cc,ck,cl,cr,cs,ct","dr","ed,eg,el,er,es,ex","f","gl,gm,gn,gr","ig,il,ir","j","kn,kr","nc,nd,nk,np,nt,ns","ocu,od,ol,oo,or","pl,pr","rd,rs,rt","sp,st","tb,tc,tt","ub,ul","w","z"].join(","),a=[c,c.toUpperCase()].join(",").split(","),u={d:"đ",D:"Đ"},l=[...r].map(e=>[[e[0],e[1]],[e[2],e[1]],[e[0].toUpperCase(),e[1].toUpperCase()],[e[2].toUpperCase(),e[1].toUpperCase()]]).reduce((e,t)=>[...e,...t],[]),f=e([["uo","ươ"],["uô","ươ"],["uó","ướ"],["uố","ướ"],["uò","ườ"],["uồ","ườ"],["uỏ","ưở"],["uổ","ưở"],["uõ","ưỡ"],["uỗ","ưỡ"],["uọ","ượ"],["uộ","ượ"]].concat([...r].map(e=>[[e[6],e[8]],[e[7],e[8]],[e[9],e[10]]]).reduce((e,t)=>[...e,...t],[]))),p=e([["uo","uô"],["ươ","uô"],["uó","uố"],["ướ","uố"],["uò","uồ"],["ườ","uồ"],["uỏ","uổ"],["ưở","uổ"],["uõ","uỗ"],["ưỡ","uỗ"],["uọ","uộ"],["ượ","uộ"]].concat([...r].map(e=>[[e[0],e[2]],[e[1],e[2]],[e[3],e[4]],[e[6],e[7]],[e[8],e[7]]]).reduce((e,t)=>[...e,...t],[]))),d=(m=[...r].map(e=>e.split("")),Object.keys(m[0]).map(e=>m.map(t=>t[e])));var m;const g={1:e([...d].map(e=>[[...e.slice(0,1),...e.slice(2)].join(""),e[1]])),2:e([...d].map(e=>[[...e.slice(0,2),...e.slice(3)].join(""),e[2]])),3:e([...d].map(e=>[[...e.slice(0,3),...e.slice(4)].join(""),e[3]])),4:e([...d].map(e=>[[...e.slice(0,4),...e.slice(5)].join(""),e[4]])),5:e([...d].map(e=>[[...e.slice(0,5)].join(""),e[5]]))},j=e=>e&&-1!==o.indexOf(e),h=e=>e&&-1!==s.indexOf(e),b=e=>e&&-1!==i.indexOf(e),k=e=>{return(t=e)&&-1!==n.indexOf(t)||j(e);var t},C=(e,t)=>{const n=g[t];for(const t of n)if(t[0].includes(e))return t[1];return e},N=(e,t,n)=>{const{count:r,mark:s,endWithVowel:i}=((e=[])=>{const t={count:0,mark:0};for(const n of e)j(n)&&t.count++,h(n)&&t.mark++;const n=e[e.length-1];return t.endWithVowel=o.indexOf(n)>=0,t})(t),c=Math.min(n,t.length);if(s>0)for(let n=c-1;n>=0;n--){const r=t[n];if(h(r)){const o=t[n+1];if(b(o))break;const s=C(r,e);if(s!=r)return t[n]=s,t.join("")}}else if(1===r)for(let n=c-1;n>=0;n--){const r=t[n];if(j(r)){const o=C(r,e);if(o!=r)return t[n]=o,t.join("")}}else if(2===r){const n=t.join("");if(!i||n.startsWith("qu")||n.startsWith("gi"))for(let n=c-1;n>=0;n--){const r=t[n];if(j(r)){const o=t[n+1];if(b(o))break;const s=C(r,e);if(s!=r)return t[n]=s,t.join("")}}else for(let n=0;n<c;n++){const r=t[n];if(j(r)){const o=t[n-1];if(b(o))break;const s=C(r,e);if(s!=r)return t[n]=s,t.join("")}}}else if(3===r){let n=!1;for(let r=c-1;r>=0;r--){const o=t[r];if(j(o)){if(!n){n=!0;continue}const s=C(o,e);if(s!=o)return t[r]=s,t.join("")}}}return t.splice(n,0,e),t.join("")},v=(e,t,n)=>{const r=t.split("");if(0===n||t.length>7||(e=>a.some(t=>e.search(t)>-1))(t))r.splice(n,0,e);else{if(9===e)return((e,t,n)=>(t[0]in u?t[0]=u[t[0]]:t[n-1]in u?t[n-1]=u[t[n-1]]:t.splice(n,0,e),t.join("")))(e,r,n);if([8,7,6].includes(e))return((e,t,n)=>{if(8===e)for(const e of l){const r=e[0],o=e[1];for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}else if(7===e)for(const e of f){const r=e[0],o=e[1];if(2===r.length){const e=t.join("");if(e.search(r)>-1)return e.replace(r,o)}else for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}else if(6===e)for(const e of p){const r=e[0],o=e[1];if(2===r.length){const e=t.join("");if(e.search(r)>-1)return e.replace(r,o)}else for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}return t.splice(n,0,e),t.join("")})(e,r,n);if([5,4,3,2,1].includes(e))return N(e,r,n)}return r.join("")},O=(e,t,n)=>{const r=t.substring(0,n),o=t.substring(n);let s=r.length;for(let e=s;e>=0;e--){const t=r[e];if(t&&!k(t))break;s=e}let i=0;for(let e=i;e<o.length;e++){const t=o[e];if(t&&!k(t))break;i=e+1}const c=n-s;i+=n;const a=[r,o].join(""),u=a.substr(0,s),l=a.substr(i,a.length),f=a.substring(s,i),p=v(e,f,c);return{replacement:[u,...p,l].join(""),curpos:n+p.length-f.length}},w=e=>"BODY"!==e.nodeName&&e.isContentEditable,y=e=>"BODY"===e.nodeName&&e.parentNode&&e.parentNode.parentNode&&"on"===e.parentNode.parentNode.designMode,E=e=>(e.cancelBubble=!0,e.stopPropagation&&e.stopPropagation(),e.preventDefault&&e.preventDefault(),!1),U=(e,t)=>{const n=(e=>{const t=e.target,n=w(t)?window.getSelection():(()=>{const e=t.parentNode.parentNode;return(e.defaultView||e.parentWindow).getSelection()})();return n&&n.rangeCount>0?n:null})(e);if(!n)return!1;const r=n.getRangeAt(0),o=r.startOffset;if(o!=r.endOffset)return!1;const s=r.startContainer,{replacement:i,curpos:c}=O(t,s.data,o);return((e,t,n,r,o)=>{const s=r.data;r.deleteData(0,s.length),r.insertData(0,t),o.setStart(r,n),E(e)})(e,i,c,s,r)},x=(e,t)=>{if(t<49||t>57)return!1;const n=t-48,r=e.target;if(y(r)||w(r))return U(e,n);const o="selectionStart"in(s=e.target)&&document.activeElement===s?s.selectionStart:-1;var s;if(-1===o)return!1;const{replacement:i,curpos:c}=O(n,r.value,o);return((e,t,n)=>{const r=e.target;r.value=t,r.focus(),r.selectionEnd=n,E(e)})(e,i,c)},D=e=>{const t=e.target,n=e.which||e.keyCode;return!!((e=>{const t=e.getAttribute("type");return"INPUT"===e.nodeName&&("text"===t||"search"===t)})(r=t)||(e=>"TEXTAREA"===e.nodeName)(r)||y(r)||w(r))&&x(e,n);var r},S=(e,t,n)=>e.addEventListener(t,n,!1),W=()=>{S(document,"keypress",D),Array.from(document.getElementsByTagName("iframe")).filter(e=>""===e.src).forEach(e=>{const t=e.contentWindow.document;S(t,"keypress",D)})};"loading"!==document.readyState?setTimeout(W,0):document.addEventListener("DOMContentLoaded",W)}();
+!function(){const e=e=>"BODY"!==e.nodeName&&e.isContentEditable,t=e=>"BODY"===e.nodeName&&e.parentNode&&e.parentNode.parentNode&&"on"===e.parentNode.parentNode.designMode,n=e=>e.map(e=>[[...e],[...e.map(e=>e.toUpperCase())]]).reduce((e,t)=>[...e,...t],[]),r="bcdđhgklmnpqrstvx",o=[r,r.toUpperCase()].join(""),s=["aăâeêioôơuưy","áắấéếíóốớúứý","àằầèềìòồờùừỳ","ảẳẩẻểỉỏổởủửỷ","ãẵẫẽễĩõỗỡũữỹ","ạặậẹệịọộợụựỵ"],i=[...s].map(e=>e+e.toUpperCase()).map(e=>e.split("")).reduce((e,t)=>[...e,...t],[]).join(""),c=[...s].map(e=>[e[1],e[2],e[4],e[7],e[8],e[10]]).reduce((e,t)=>[...e,...t],[]).join(""),a=[...s].slice(1).join(""),u=["ab,ace,act,ad,ae,ann,al,ag,ar,as","bb,br,bl","cc,ck,cl,cr,cs,ct","dr","ed,eg,el,er,es,ex","f","gl,gm,gn,gr","ig,il,ir","j","kn,kr","nc,nd,nk,np,nt,ns","ocu,od,ol,oo,or","pl,pr","rd,rs,rt","sp,st","tb,tc,tt","ub,ul","w","z"].join(","),l=[u,u.toUpperCase()].join(",").split(","),f={d:"đ",D:"Đ"},p=[...s].map(e=>[[e[0],e[1]],[e[2],e[1]],[e[0].toUpperCase(),e[1].toUpperCase()],[e[2].toUpperCase(),e[1].toUpperCase()]]).reduce((e,t)=>[...e,...t],[]),d=n([["uo","ươ"],["uô","ươ"],["uó","ướ"],["uố","ướ"],["uò","ườ"],["uồ","ườ"],["uỏ","ưở"],["uổ","ưở"],["uõ","ưỡ"],["uỗ","ưỡ"],["uọ","ượ"],["uộ","ượ"]].concat([...s].map(e=>[[e[6],e[8]],[e[7],e[8]],[e[9],e[10]]]).reduce((e,t)=>[...e,...t],[]))),m=n([["uo","uô"],["ươ","uô"],["uó","uố"],["ướ","uố"],["uò","uồ"],["ườ","uồ"],["uỏ","uổ"],["ưở","uổ"],["uõ","uỗ"],["ưỡ","uỗ"],["uọ","uộ"],["ượ","uộ"]].concat([...s].map(e=>[[e[0],e[2]],[e[1],e[2]],[e[3],e[4]],[e[6],e[7]],[e[8],e[7]]]).reduce((e,t)=>[...e,...t],[]))),g=(j=[...s].map(e=>e.split("")),Object.keys(j[0]).map(e=>j.map(t=>t[e])));var j;const h={1:n([...g].map(e=>[[...e.slice(0,1),...e.slice(2)].join(""),e[1]])),2:n([...g].map(e=>[[...e.slice(0,2),...e.slice(3)].join(""),e[2]])),3:n([...g].map(e=>[[...e.slice(0,3),...e.slice(4)].join(""),e[3]])),4:n([...g].map(e=>[[...e.slice(0,4),...e.slice(5)].join(""),e[4]])),5:n([...g].map(e=>[[...e.slice(0,5)].join(""),e[5]]))},b=e=>e&&-1!==i.indexOf(e),k=e=>e&&-1!==c.indexOf(e),C=e=>e&&-1!==a.indexOf(e),N=e=>{return(t=e)&&-1!==o.indexOf(t)||b(e);var t},v=(e,t)=>{const n=h[t];for(const t of n)if(t[0].includes(e))return t[1];return e},O=(e,t,n)=>{const{count:r,mark:o,endWithVowel:s}=((e=[])=>{const t={count:0,mark:0};for(const n of e)b(n)&&t.count++,k(n)&&t.mark++;const n=e[e.length-1];return t.endWithVowel=i.indexOf(n)>=0,t})(t),c=Math.min(n,t.length);if(o>0)for(let n=c-1;n>=0;n--){const r=t[n];if(k(r)){const o=t[n+1];if(C(o))break;const s=v(r,e);if(s!=r)return t[n]=s,t.join("")}}else if(1===r)for(let n=c-1;n>=0;n--){const r=t[n];if(b(r)){const o=v(r,e);if(o!=r)return t[n]=o,t.join("")}}else if(2===r){const n=t.join("");if(!s||n.startsWith("qu")||n.startsWith("gi"))for(let n=c-1;n>=0;n--){const r=t[n];if(b(r)){const o=t[n+1];if(C(o))break;const s=v(r,e);if(s!=r)return t[n]=s,t.join("")}}else for(let n=0;n<c;n++){const r=t[n];if(b(r)){const o=t[n-1];if(C(o))break;const s=v(r,e);if(s!=r)return t[n]=s,t.join("")}}}else if(3===r){let n=!1;for(let r=c-1;r>=0;r--){const o=t[r];if(b(o)){if(!n){n=!0;continue}const s=v(o,e);if(s!=o)return t[r]=s,t.join("")}}}return t.splice(n,0,e),t.join("")},w=(e,t,n)=>{const r=t.split("");if(0===n||t.length>7||(e=>l.some(t=>e.search(t)>-1))(t))r.splice(n,0,e);else{if(9===e)return((e,t,n)=>(t[0]in f?t[0]=f[t[0]]:t[n-1]in f?t[n-1]=f[t[n-1]]:t.splice(n,0,e),t.join("")))(e,r,n);if([8,7,6].includes(e))return((e,t,n)=>{if(8===e)for(const e of p){const r=e[0],o=e[1];for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}else if(7===e)for(const e of d){const r=e[0],o=e[1];if(2===r.length){const e=t.join("");if(e.search(r)>-1)return e.replace(r,o)}else for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}else if(6===e)for(const e of m){const r=e[0],o=e[1];if(2===r.length){const e=t.join("");if(e.search(r)>-1)return e.replace(r,o)}else for(let e=n-1;e>=0;e--){if(t[e]===r)return t[e]=o,t.join("")}}return t.splice(n,0,e),t.join("")})(e,r,n);if([5,4,3,2,1].includes(e))return O(e,r,n)}return r.join("")},x=(e,t,n)=>{const r=t.substring(0,n),o=t.substring(n);let s=r.length;for(let e=s;e>=0;e--){const t=r[e];if(t&&!N(t))break;s=e}let i=0;for(let e=i;e<o.length;e++){const t=o[e];if(t&&!N(t))break;i=e+1}const c=n-s;i+=n;const a=[r,o].join(""),u=a.substr(0,s),l=a.substr(i,a.length),f=a.substring(s,i),p=w(e,f,c);return{replacement:[u,...p,l].join(""),curpos:n+p.length-f.length}},y=e=>(e.cancelBubble=!0,e.stopPropagation&&e.stopPropagation(),e.preventDefault&&e.preventDefault(),!1),E=(t,n)=>{const r=(t=>{const n=t.target,r=e(n)?window.getSelection():(()=>{const e=n.parentNode.parentNode;return(e.defaultView||e.parentWindow).getSelection()})();return r&&r.rangeCount>0?r:null})(t);if(!r)return!1;const o=r.getRangeAt(0),s=o.startOffset;if(s!=o.endOffset)return!1;const i=o.startContainer,{replacement:c,curpos:a}=x(n,i.data,s);return((e,t,n,r,o)=>{const s=r.data;r.deleteData(0,s.length),r.insertData(0,t),o.setStart(r,n),y(e)})(t,c,a,i,o)},U=(n,r)=>{if(r<49||r>57)return!1;const o=r-48,s=n.target;if(t(s)||e(s))return E(n,o);const i="selectionStart"in(c=n.target)&&document.activeElement===c?c.selectionStart:-1;var c;if(-1===i)return!1;const{replacement:a,curpos:u}=x(o,s.value,i);return((e,t,n)=>{const r=e.target;r.value=t,r.focus(),r.selectionEnd=n,y(e)})(n,a,u)},D=n=>{const r=n.target,o=n.which||n.keyCode;return!!((e=>{const t=e.getAttribute("type")||"text";return"INPUT"===e.nodeName&&("text"===t||"search"===t)})(s=r)||(e=>"TEXTAREA"===e.nodeName)(s)||t(s)||e(s))&&U(n,o);var s},S=(e,t,n)=>e.addEventListener(t,n,!1),W=()=>{S(document,"keypress",D),Array.from(document.getElementsByTagName("iframe")).filter(e=>""===e.src).forEach(e=>{const t=e.contentWindow.document;S(t,"keypress",D)})};"loading"!==document.readyState?setTimeout(W,0):document.addEventListener("DOMContentLoaded",W)}();
