+++ docs/index.html
@@ -1,5 +1,5 @@
-<html><head> <meta charset="utf-8"> <meta name="viewport"content="width=device-width,initial-scale=1"> <title>vnikey - Demo</title> <link rel="subresource"href="style.css?rev=6jyLMtsCwZ6hlS5z40ClarVH"><link rel="stylesheet"type="text/css"href="style.css?rev=6jyLMtsCwZ6hlS5z40ClarVH"></head> <body> <header> <h1><a href="https://github.com/ndaidong/vnikey">vnikey</a></h1> </header> <div class="container"> <label> Text input:</label> <input type="text"> <label> Text area:</label> <textarea class="txtbox"id="txtbox"></textarea> <label> Contenteditable element:</label> <pre id="editor"contenteditable></pre> <label> Iframe with designMode enabled:</label> <iframe id="iframeEditor"></iframe> </div> <footer> <hr> <div class="copyright"> <a href="https://twitter.com/@ndaidong"target="twitter">@ndaidong</a> </div> </footer> <script> const editor = document.getElementById('iframeEditor').contentWindow.document;
+<html><head> <meta charset="utf-8"> <meta name="viewport"content="width=device-width,initial-scale=1"> <title>vnikey - Demo</title> <link rel="subresource"href="style.css?rev=kNtiZeTzDzGZd2rrocdkdc5c"><link rel="stylesheet"type="text/css"href="style.css?rev=kNtiZeTzDzGZd2rrocdkdc5c"></head> <body> <header> <h1><a href="https://github.com/ndaidong/vnikey">vnikey</a></h1> </header> <div class="container"> <label> Text input:</label> <input type="text"> <label> Text area:</label> <textarea class="txtbox"id="txtbox"></textarea> <label> Contenteditable element:</label> <pre id="editor"contenteditable></pre> <label> Iframe with designMode enabled:</label> <iframe id="iframeEditor"></iframe> </div> <footer> <hr> <div class="copyright"> <a href="https://twitter.com/@ndaidong"target="twitter">@ndaidong</a> </div> </footer> <script> const editor = document.getElementById('iframeEditor').contentWindow.document;
       editor.designMode = 'On';
       editor.open();
       editor.write('<html><head></head><body style="font-size: 16px; background-color: #000; color: #fff;"></body></html>');
-      editor.close(); </script> <script type="text/javascript"src="app.js?rev=6jyLMtsCwZ6hlS5z40ClarVH"></script></body></html>
+      editor.close(); </script> <script type="text/javascript"src="app.js?rev=kNtiZeTzDzGZd2rrocdkdc5c"></script></body></html>
