+++ extensions/webkit/manifest.json
@@ -3,7 +3,7 @@
   "name": "vnikey for Webkit",
   "description": "Enables Vietnamese number key based",
   "author": "@ndaidong",
-  "version": "0.0.9",
+  "version": "0.2.0",
   "homepage_url": "https://github.com/ndaidong/vnikey",
   "icons": {
     "16": "icons/vni-16x16.png",
